+++ src/aten/Indexing.cpp
@@ -1,3 +1,4 @@
+#include <ATen/ATen.h>
 #include <ATen/XPUNativeFunctions.h>
 #include <aten/sycl/Indexing.h>
 
@@ -8,14 +9,7 @@ Tensor& XPUNativeFunctions::index_select_out(
     int64_t dim,
     const Tensor& index,
     Tensor& out) {
-  AT_DISPATCH_ALL_TYPES_AND_COMPLEX_AND3(
-      at::ScalarType::Half,
-      at::ScalarType::BFloat16,
-      at::ScalarType::Bool,
-      self.scalar_type(),
-      "indexSelect",
-      [=]() { native::xpu::indexSelect<scalar_t>(out, self, dim, index); });
-  return out;
+  return at::native::xpu::index_select_out_kernel(self, dim, index, out);
 }
 
 Tensor XPUNativeFunctions::index_select(
@@ -23,6 +17,6 @@ Tensor XPUNativeFunctions::index_select(
     int64_t dim,
     const Tensor& index) {
   auto out = at::empty({0}, self.options());
-  return index_select_out(self, dim, index, out);
+  return at::native::xpu::index_select_out_kernel(self, dim, index, out);
 }
 } // namespace at
